RefreshToken 엔티티 생성 및 Student 엔티티, DTO 타입 수정

+++ src/main/java/com/example/quick_gg/repository/StudentRepository.java
@@ -20,6 +20,6 @@ public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
          boolean existsByStudentNumber(String studentNumber);
 
          // id가 이미 등록되어 있는지 확인
-         boolean existsById(Long id);
+         boolean existsById(Integer id);
 
 }
